Another semi-major change
Due to issues with how sonarr manages the "clean" version of naming, I decided to switch both sonarr/radarr to the "Unclean" version. This all around allows for more consistency. The main reason is it appears to be inconsistent on what is removed and what isn't when using the "Clean" version.