Make backup_appdata work if container needs file binds
In rare cases a container image requires binding a file directly, which the script takes as source_dir for the backup. This adds a check, if the directory the scripts grabs is a filename and strips the filename.

i.e. ./backup_appdata.sh: line 185: cd: /appdata/container/container.conf/..: Not a directory stripes the filename (container.conf), so the script backups /appdata/container/ and its content properly.