Added a JSON metadata check for asset folders
That way it doesn't need to iterate over all directories multiple times, I noticed a slow-down here, so I implemented this to prevent that. For now will only work on my machine since I am the only one with such a json file.

Feel free to standardize this into the libraries themselves.